+++ src/main/java/com/tourranger/travelAgency/repository/TravelAgencyRepository.java
@@ -1,9 +1,9 @@
-package travelAgency.repository;
+package com.tourranger.travelAgency.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import travelAgency.entity.TravelAgency;
+import com.tourranger.travelAgency.entity.TravelAgency;
 
 @Repository
 public interface TravelAgencyRepository extends JpaRepository<TravelAgency, Long>, TravelAgencyRepositoryCustom {
