+++ src/main/java/com/tourranger/travelAgency/repository/TravelAgencyRepositoryCustom.java
@@ -1,4 +1,4 @@
-package travelAgency.repository;
+package com.tourranger.travelAgency.repository;
 
 public interface TravelAgencyRepositoryCustom {
