+++ src/main/java/com/tourranger/travelAgency/service/TravelAgencyServiceImpl.java
@@ -1,4 +1,4 @@
-package com.tourranger.thumbnailImage.service;
+package com.tourranger.travelAgency.service;
 
 import java.net.MalformedURLException;
 
@@ -10,29 +10,30 @@ import org.springframework.transaction.annotation.Transactional;
 import com.tourranger.common.error.CustomErrorCode;
 import com.tourranger.common.exception.CustomException;
 import com.tourranger.common.file.FileStore;
-import com.tourranger.thumbnailImage.entity.ThumbnailImage;
-import com.tourranger.thumbnailImage.repository.ThumbnailImageRepository;
+import com.tourranger.travelAgency.entity.TravelAgency;
 
 import lombok.RequiredArgsConstructor;
 
+import com.tourranger.travelAgency.repository.TravelAgencyRepository;
+
 @Service
 @RequiredArgsConstructor
-public class ThumbnailImageServiceImpl implements ThumbnailImageService {
-	private final ThumbnailImageRepository thumbnailImageRepository;
+public class TravelAgencyServiceImpl implements TravelAgencyService {
+	private final TravelAgencyRepository travelAgencyRepository;
 	private final FileStore fileStore;
 
 	//이미지를 조회하는 메서드
 	@Override
 	@Transactional(readOnly = true)
-	public Resource getThumbnailImage(Long thumbnailImageId) throws MalformedURLException {
-		ThumbnailImage thumbnailImage = findThumbnailImage(thumbnailImageId);
-		return new UrlResource(thumbnailImage.getUrl());
+	public Resource getTravelAgency(Long travelAgencyId) throws MalformedURLException {
+		TravelAgency travelAgency = findTravelAgency(travelAgencyId);
+		return new UrlResource(travelAgency.getUrl());
 	}
 
 	// 이미지 객체를 repository에서 찾는 메서드
 	// id로 조회했을 때, 존재하지 않는 이미지인 경우 Exception 발생
-	private ThumbnailImage findThumbnailImage(Long thumbnailImageId) {
-		return thumbnailImageRepository.findById(thumbnailImageId).orElseThrow(() ->
+	private TravelAgency findTravelAgency(Long travelAgencyId) {
+		return travelAgencyRepository.findById(travelAgencyId).orElseThrow(() ->
 			new CustomException(CustomErrorCode.IMAGE_NOT_FOUND, null)
 
 		);
